requirements.txt
annotated-types==0.7.0
antiword==0.1.0
argcomplete==1.10.0
beautifulsoup4==4.8.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.2
olefile==0.46
packaging==24.2
pandas==2.2.3
pdfminer==20191125
pdfminer.six==20181108
pillow==11.1.0
preshed==3.0.9
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-pptx==0.6.18
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==1.5.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.2
